Organize code - moved all map components inside Map Directory
